backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
google-genai
python-multipart
httpx
firebase-admin
pydantic
